FAD-5559 Rename folders & filenames
