+++ src/pages/billing/forms/fields/BillingContactForm.js
@@ -9,7 +9,7 @@ import { TextFieldWrapper, SelectWrapper } from 'src/components';
 import { required, email } from 'src/helpers/validation';
 import { getZipLabel } from 'src/helpers/billing';
 
-import styles from './Forms.module.scss';
+import styles from './Fields.module.scss';
 
 /**
  * This component will register the following redux-form fields
